pyproject.toml
[tool.poetry]
name = "medicaltextautocomplete"
version = "0.1.0"
description = "Final qualifying work on the topic 'Automatic addition of medical texts using a pre-trained artificial intelligence model'"
authors = ["Игорь Иванов <is.ivanov@bars.group>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.2.0"
transformers = "^4.37.2"
sacremoses = "^0.1.1"
psycopg2-binary = "^2.9.9"
pyyaml = "^6.0.1"
click = "^8.1.7"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
tqdm = "^4.66.2"
selenium = "^4.18.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
get_data = "src.data.get_data:process"